chore: renamed main circuit to zkube circuit
