--- circuits/circuits/main.json
@@ -1,40 +1,0 @@
-{
-  "initialGrid": [
-    [1, 1, 2, 0, 0, 0, 0, 0],
-    [1, 2, 0, 0, 0, 0, 0, 0],
-    [2, 0, 0, 0, 0, 0, 0, 0],
-    [2, 0, 0, 0, 0, 0, 0, 0],
-    [1, 2, 0, 0, 0, 0, 0, 0],
-    [1, 1, 2, 0, 0, 0, 0, 0],
-    [0, 0, 0, 0, 0, 0, 0, 0],
-    [0, 0, 0, 0, 0, 0, 0, 0]
-  ],
-  "finalGrid": [
-    [2, 2, 2, 2, 0, 0, 0, 0],
-    [2, 2, 2, 0, 0, 0, 0, 0],
-    [2, 2, 0, 0, 0, 0, 0, 0],
-    [2, 2, 0, 0, 0, 0, 0, 0],
-    [2, 2, 2, 0, 0, 0, 0, 0],
-    [2, 2, 2, 2, 0, 0, 0, 0],
-    [0, 0, 0, 0, 0, 0, 0, 0],
-    [0, 0, 0, 0, 0, 0, 0, 0]
-  ],
-  "account": "0x123",
-  "selectedFunctions": [
-    [
-      [0, 0, 0],
-      [1, 2, 0],
-      [0, 0, 0]
-    ],
-    [
-      [1, 1, 2],
-      [0, 0, 0],
-      [0, 0, 0]
-    ],
-    [
-      [0, 0, 0],
-      [0, 0, 0],
-      [0, 0, 0]
-    ]
-  ]
-}
