+++ circuits/zk/input.json
@@ -1,0 +1,1 @@
+{"initialGrid":[[1,1,2,0,0,0,0,0],[1,2,0,0,0,0,0,0],[2,0,0,0,0,0,0,0],[2,0,0,0,0,0,0,0],[1,2,0,0,0,0,0,0],[1,1,2,0,0,0,0,0],[0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0]],"finalGrid":[[2,3,2,3,2,3,2,3],[2,3,2,3,2,3,0,0],[2,3,2,3,0,0,0,0],[2,3,2,3,0,0,0,0],[2,3,2,3,2,3,0,0],[2,3,2,3,2,3,2,3],[0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0]],"account":"0x123","selectedFunctionsIndexes":[1,8,13]}
