+++ circuits/zk/proof.json
@@ -1,0 +1,28 @@
+{
+ "pi_a": [
+  "13379666701580020387694903898194761326852962271926430641493962382595303440518",
+  "5673969288966157796196460184889918633482645929288461669935990585323575711956",
+  "1"
+ ],
+ "pi_b": [
+  [
+   "10410398817799012434999496471248576087893757059143322230014935871366329368341",
+   "2336425364646520336587791655568623717321048254070738925744156771200413140072"
+  ],
+  [
+   "10995600481687705778706248275130102690382342484466737630530112445816334752419",
+   "2099279396996001841797025437617838911529012718826108633893335597482527255306"
+  ],
+  [
+   "1",
+   "0"
+  ]
+ ],
+ "pi_c": [
+  "1926950336407856473186440297771789333924126449741864350630525192393450294060",
+  "331422420395719853209066067677110454588473793055120679087121090107748371796",
+  "1"
+ ],
+ "protocol": "groth16",
+ "curve": "bn128"
+}
