+++ circuits/zk/public.json
@@ -1,0 +1,131 @@
+[
+ "1",
+ "1",
+ "2",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "1",
+ "2",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "2",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "2",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "1",
+ "2",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "1",
+ "1",
+ "2",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "2",
+ "3",
+ "2",
+ "3",
+ "2",
+ "3",
+ "2",
+ "3",
+ "2",
+ "3",
+ "2",
+ "3",
+ "2",
+ "3",
+ "0",
+ "0",
+ "2",
+ "3",
+ "2",
+ "3",
+ "0",
+ "0",
+ "0",
+ "0",
+ "2",
+ "3",
+ "2",
+ "3",
+ "0",
+ "0",
+ "0",
+ "0",
+ "2",
+ "3",
+ "2",
+ "3",
+ "2",
+ "3",
+ "0",
+ "0",
+ "2",
+ "3",
+ "2",
+ "3",
+ "2",
+ "3",
+ "2",
+ "3",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "0",
+ "291"
+]
